refactor: move getInitials into text utils
