--- src/utils.ts
@@ -1,13 +1,0 @@
-export function getInitials(name: string | null | undefined): string {
-  if (!name) {
-    return "";
-  }
-
-  const parts = name.trim().split(/\s+/);
-
-  if (parts.length === 1) {
-    return parts[0][0]?.toUpperCase() ?? "";
-  }
-
-  return (parts[0][0] + parts[parts.length - 1][0]).toUpperCase();
-}
