+++ src/utils/text.ts
@@ -1,2 +1,16 @@
 export const trimLeadingSlashes = (str: string) => str.replace(/^\/+/g, "");
 export const trimTrailingSlashes = (str: string) => str.replace(/\/+$/g, "");
+
+export function getInitials(name: string | null | undefined): string {
+  if (!name) {
+    return "";
+  }
+
+  const parts = name.trim().split(/\s+/);
+
+  if (parts.length === 1) {
+    return parts[0][0]?.toUpperCase() ?? "";
+  }
+
+  return (parts[0][0] + parts[parts.length - 1][0]).toUpperCase();
+}
